requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai>=1.0.0
Pillow==10.2.0
gunicorn==21.2.0
numpy==1.26.4
scipy==1.13.0
fastdtw==0.3.4
